File relocation
- created separate packages for:
1) dtos
2) custom exceptions
